+++ src/main/java/com/example/demo/configuration/BeanDefinitions.java
@@ -1,5 +1,6 @@
-package com.example.demo.service.configuration;
+package com.example.demo.configuration;
 
+import ca.uhn.fhir.context.FhirContext;
 import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
@@ -10,5 +11,10 @@ public class BeanDefinitions {
     @Bean
     public RestTemplate restTemplate() {
         return new RestTemplate();
+    }
+
+    @Bean
+    public FhirContext getFhirContext() {
+        return FhirContext.forR4();
     }
 }
