+++ src/main/java/com/example/demo/model/Person.java
@@ -1,4 +1,4 @@
-package com.example.demo.entities;
+package com.example.demo.model;
 
 public class Person {
     String firstName;
