--- src/main/java/com/example/demo/service/configuration/LocalFhirContext.java
@@ -1,15 +1,0 @@
-package com.example.demo.service.configuration;
-
-import ca.uhn.fhir.context.FhirContext;
-import org.springframework.stereotype.Component;
-
-@Component
-public class LocalFhirContext {
-    // Erzeugt ein FhirContext-Objekt für FHIR R4
-
-    private final FhirContext fhirContext = FhirContext.forR4();
-
-    public FhirContext getFhirContext() {
-        return fhirContext;
-    }
-}
